create notifications repository
